[AIRFLOW-3646] Rename plugins_manager.py to test_plugins_manager.py to trigger tests
