+++ tests/plugins/test_plugins_manager_www.py
@@ -27,13 +27,11 @@ import unittest
 from flask.blueprints import Blueprint
 from flask_admin.menu import MenuLink, MenuView
 
-from airflow.configuration import conf
 from airflow.hooks.base_hook import BaseHook
 from airflow.models import BaseOperator
 from airflow.sensors.base_sensor_operator import BaseSensorOperator
 from airflow.executors.base_executor import BaseExecutor
-from airflow.www.app import cached_app
-from airflow.www_rbac import app as application
+from airflow.www.app import create_app
 
 
 class PluginsTest(unittest.TestCase):
@@ -67,7 +65,7 @@ class PluginsTest(unittest.TestCase):
         self.assertTrue(callable(plugin_macro))
 
     def test_admin_views(self):
-        app = cached_app()
+        app = create_app(testing=True)
         [admin] = app.extensions['admin']
         category = admin._menu_categories['Test Plugin']
         [admin_view] = [v for v in category.get_children()
@@ -75,52 +73,13 @@ class PluginsTest(unittest.TestCase):
         self.assertEqual('Test View', admin_view.name)
 
     def test_flask_blueprints(self):
-        app = cached_app()
+        app = create_app(testing=True)
         self.assertIsInstance(app.blueprints['test_plugin'], Blueprint)
 
     def test_menu_links(self):
-        app = cached_app()
+        app = create_app(testing=True)
         [admin] = app.extensions['admin']
         category = admin._menu_categories['Test Plugin']
         [menu_link] = [ml for ml in category.get_children()
                        if isinstance(ml, MenuLink)]
         self.assertEqual('Test Menu Link', menu_link.name)
-
-
-class PluginsTestRBAC(unittest.TestCase):
-    def setUp(self):
-        conf.load_test_config()
-        self.app, self.appbuilder = application.create_app(testing=True)
-
-    def test_flaskappbuilder_views(self):
-        from tests.plugins.test_plugin import v_appbuilder_package
-        appbuilder_class_name = str(v_appbuilder_package['view'].__class__.__name__)
-        plugin_views = [view for view in self.appbuilder.baseviews
-                        if view.blueprint.name == appbuilder_class_name]
-
-        self.assertTrue(len(plugin_views) == 1)
-
-        # view should have a menu item matching category of v_appbuilder_package
-        links = [menu_item for menu_item in self.appbuilder.menu.menu
-                 if menu_item.name == v_appbuilder_package['category']]
-
-        self.assertTrue(len(links) == 1)
-
-        # menu link should also have a link matching the name of the package.
-        link = links[0]
-        self.assertEqual(link.name, v_appbuilder_package['category'])
-        self.assertEqual(link.childs[0].name, v_appbuilder_package['name'])
-
-    def test_flaskappbuilder_menu_links(self):
-        from tests.plugins.test_plugin import appbuilder_mitem
-
-        # menu item should exist matching appbuilder_mitem
-        links = [menu_item for menu_item in self.appbuilder.menu.menu
-                 if menu_item.name == appbuilder_mitem['category']]
-
-        self.assertTrue(len(links) == 1)
-
-        # menu link should also have a link matching the name of the package.
-        link = links[0]
-        self.assertEqual(link.name, appbuilder_mitem['category'])
-        self.assertEqual(link.childs[0].name, appbuilder_mitem['name'])
